// src/app/shared/enums/message.ts
export enum MessageVisibility {
    TO_ALL = 'TO_ALL',
    ONLY_TO_ME = 'ONLY_TO_ME',
    TO_NO_ONE = 'TO_NO_ONE',
    TO_AGENTS = 'TO_AGENTS'
};



export enum MessageVisibilityIcons {
    DEFAULT = 'more_vert',
    TO_ALL = 'visibility',
    ONLY_TO_ME = 'visibility_off',
    TO_NO_ONE = 'delete',
    TO_AGENTS = 'disabled_visible'
}



export enum MessageAttachmentType {
    IMAGE = 'IMAGE',
    VIDEO = 'VIDEO',
    TEXT = 'TEXT',
    PDF = 'PDF',
    OTHER = 'OTHER',
};